start converting to the epics format
